rename file: report-html-generator-v2 -> report-html-generator-impl
