+++ src/DetailsView/reports/report-html-generator-impl.tsx
